requirements.txt
nose
pyahocorasick
spacy==0.101.0
elasticsearch==2.3.0
numpy==1.11.0
PyYAML==3.11
urllib3==1.16

## scraper
parsedatetime
bs4
APScheduler
boto3
cssselect
futures
goose-extractor
jieba
jusText
lxml
nltk
Pattern
Pillow
pymongo
pytz
requests
six
tzlocal
